move builder and artifacts to modles
